pyproject.toml
[build-system]
requires = ["setuptools >= 61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "hf-security-scanner"
version = "0.1.0"
description = "Comprehensive security auditing tool for Hugging Face models"
readme = "README.md"
requires-python = ">=3.8"
authors = [
    {name = "Mayukh Dey", email = "mactash076@gmail.com"}
]
keywords = ["huggingface", "security", "model-scanning", "ai-security", "ml-security"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
]
dependencies = [
    "requests>=2.28.0",
    "urllib3>=1.26.0",
    "tqdm>=4.64.0",
    "click>=8.1.0",
    "pyyaml>=6.0",
    "colorlog>=6.7.0",
    "rich>=13.0.0",
    "huggingface-hub>=0.15.0",
    "transformers>=4.21.0",
    "safetensors>=0.3.0",
    "python-magic>=0.4.24",
    "pycryptodome>=3.17.0",
    "cryptography>=39.0.0",
    "bandit>=1.7.0",
    "pandas>=1.5.0",
    "numpy>=1.21.0",
    "sqlalchemy>=1.4.0",
    "alembic>=1.10.0",
    "jinja2>=3.1.0",
    "markdown>=3.4.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "black>=22.0.0",
    "flake8>=5.0.0",
    "mypy>=0.991",
    "pre-commit>=2.20.0",
]
docs = [
    "sphinx>=5.0.0",
    "sphinx-rtd-theme>=1.0.0",
]

[project.scripts]
hf-scan = "hf_security_scanner.cli.main:cli"

[tool.setuptools.packages.find]
where = ["src"]

[tool.black]
line-length = 88
target-version = ['py38']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[tool.mypy]
python_version = "3.8"
check_untyped_defs = true
ignore_missing_imports = true
warn_unused_ignores = true
warn_redundant_casts = true
warn_unused_configs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --tb=short --strict-markers"
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "security: marks tests as security-related",
    "network: marks tests that require network access",
]